Added wasm32 support
